# docs/contracts/control-api.openapi.yaml
openapi: 3.1.0
info:
  title: Forge Control Plane API
  version: 0.1.0
  description: |
    HTTP API for the Forge control plane. Used by:
    - appx-api (via forge-sdk-ts)
    - forge-cli
    - forge-agent (registration, heartbeats, command polling)

    All endpoints require Bearer token auth except /healthz and /metrics.
    Agent endpoints additionally require a node-scoped token.

    Error responses use application/problem+json (RFC 7807).
servers:
  - url: https://forge.internal.myappx.live/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local dev
security:
  - bearerAuth: []
paths:
...
  /sandboxes:
    post:
      summary: Create a sandbox
      operationId: createSandbox
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SandboxCreate'
      responses:
        '201':
          description: Sandbox created (in 'pending' state, scheduling async)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '409':
          description: app_name already exists
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: No nodes available
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List sandboxes
      operationId: listSandboxes
      parameters:
        - in: query
          name: app_name
          schema: { type: string }
        - in: query
          name: user_id
          schema: { type: string }
        - in: query
          name: state
          schema:
            type: string
            enum: [pending, starting, running, restarting, destroying, destroyed, failed, stopped]
        - in: query
          name: node_id
          schema: { type: string, format: uuid }
        - in: query
          name: limit
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        '200':
          description: List of sandboxes
          content:
            application/json:
              schema:
                type: object
                properties:
                  sandboxes:
                    type: array
                    items: { $ref: '#/components/schemas/Sandbox' }
  /sandboxes/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          oneOf:
            - { type: string, format: uuid, description: Sandbox UUID }
            - { type: string, pattern: '^app:.+', description: 'Lookup by app_name with prefix "app:"' }
    get:
      summary: Get sandbox
      operationId: getSandbox
      responses:
        '200':
          description: Sandbox details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sandbox'
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Destroy sandbox
      operationId: destroySandbox
      responses:
        '202':
          description: Destruction queued
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
  /sandboxes/{id}/restart:
    post:
      summary: Force restart a sandbox
      operationId: restartSandbox
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '202':
          description: Restart queued
        '404':
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    SandboxCreate:
      type: object
      required: [app_name, user_id, image]
      properties:
        app_name:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{1,62}$'
          description: Becomes {app_name}.myappx.live
        user_id: { type: string }
        image: { type: string, example: 'appx/sandbox:v1' }
        resources:
          type: object
          properties:
            cpu_cores: { type: number, default: 0.5 }
            memory_mb: { type: integer, default: 512 }
        env:
          type: object
          additionalProperties: { type: string }
        idle_timeout_seconds:
          type: integer
          default: 1800
          description: Auto-stop sandbox after this many seconds of inactivity
        metadata: { type: object }
    Sandbox:
      type: object
      properties:
        id: { type: string, format: uuid }
        app_name: { type: string }
        user_id: { type: string }
        node_id: { type: string, format: uuid, nullable: true }
        container_id: { type: string, nullable: true }
        image: { type: string }
        state:
          type: string
          enum: [pending, starting, running, restarting, destroying, destroyed, failed, stopped]
        url: { type: string, format: uri, description: 'Public URL: https://{app_name}.myappx.live' }
        resources:
          type: object
          properties:
            cpu_cores: { type: number }
            memory_mb: { type: integer }
        host_port: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_active_at: { type: string, format: date-time }
        failure_count: { type: integer }
        metadata: { type: object }
...
    Error:
      type: object
      description: RFC 7807 Problem Details
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, description: 'Problem type URI', example: 'urn:forge:error:conflict' }
        title: { type: string, description: 'Short human-readable summary', example: 'app_name already exists' }
        status: { type: integer, description: 'HTTP status code', example: 409 }
        detail: { type: string, description: 'Detailed human-readable explanation' }
        instance: { type: string, format: uri, description: 'URI identifying the specific occurrence' }